TST: switch names of test files (not sure if git tracks this)
